# api/app/openapi/OpenAPI_backend_En.yaml
openapi: 3.1.0
info:
  title: Dev Portal & Translation → Chroma Pipeline API
  version: 0.3.0
  description: >
    An API that provides the pipeline: Extract → Translate → (optional Writeback) → Packaging → Chroma upsert,
    plus CRUD/import operations for the portal_* tables.
    Reflects Phase J/K policies and clarifies POST idempotency / Request-ID / 429/409 responses.
...
servers:
  - url: https://api.example.com
    description: default (switch prod/dev via env overlay)
security: []  # In the prod overlay, apply security: [{ bearerAuth: [] }]
...
paths:
...
  /portal/model:
    get:
      tags: [Portal Model]
      summary: List models
      parameters:
        - in: query
          name: q
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer, default: 50, minimum: 1, maximum: 500 }
        - in: query
          name: cursor
          schema: { type: string }
      responses:
        '200':
          description: OK
          content: { application/json: { schema: { $ref: '#/components/schemas/PortalModelList' } } }
        default: { $ref: '#/components/responses/Problem' }
    post:
      tags: [Portal Model]
      summary: Create model
      parameters:
        - $ref: '#/components/parameters/XRequestID'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/PortalModelCreate' } } }
      responses:
        '201':
          description: Created
          content: { application/json: { schema: { $ref: '#/components/schemas/PortalModel' } } }
        '409': { $ref: '#/components/responses/ConflictIdempotent' }
        default: { $ref: '#/components/responses/Problem' }
...
  /portal/model/import:
    post:
      tags: [Portal Model]
      summary: Selective import from ir_model_src (optional scaffold)
      parameters:
        - $ref: '#/components/parameters/XRequestID'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/ImportModelRequest' } } }
      responses:
        '200':
          description: OK
          content: { application/json: { schema: { $ref: '#/components/schemas/ImportResult' } } }
        '409': { $ref: '#/components/responses/ConflictIdempotent' }
        default: { $ref: '#/components/responses/Problem' }

  ############################################
  # Portal: field
  ############################################
...
  /portal/smart_button:
    get:
      tags: [Portal Smart Button]
      summary: List smart buttons
      parameters:
        - in: query
          name: view_id
          schema: { type: integer }
        - in: query
          name: button_key
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer, default: 50 }
        - in: query
          name: cursor
          schema: { type: string }
      responses:
        '200':
          description: OK
          content: { application/json: { schema: { $ref: '#/components/schemas/PortalSmartButtonList' } } }
        default: { $ref: '#/components/responses/Problem' }
    post:
      tags: [Portal Smart Button]
      summary: Create
      parameters:
        - $ref: '#/components/parameters/XRequestID'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/PortalSmartButtonCreate' } } }
      responses:
        '201':
          description: Created
          content: { application/json: { schema: { $ref: '#/components/schemas/PortalSmartButton' } } }
        '409': { $ref: '#/components/responses/ConflictIdempotent' }
        default: { $ref: '#/components/responses/Problem' }
  /portal/smart_button/{id}:
    patch:
      tags: [Portal Smart Button]
      summary: Update (partial)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
        - $ref: '#/components/parameters/XRequestID'
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/PortalSmartButtonUpdate' } } }
      responses:
        '200':
          description: OK
          content: { application/json: { schema: { $ref: '#/components/schemas/PortalSmartButton' } } }
        default: { $ref: '#/components/responses/Problem' }
  /portal/smart_button/bulk_upsert:
    post:
      tags: [Portal Smart Button]
      summary: Bulk UPSERT multiple buttons
      parameters:
        - $ref: '#/components/parameters/XRequestID'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/BulkPortalSmartButtonRequest' } } }
      responses:
        '200':
          description: OK
          content: { application/json: { schema: { $ref: '#/components/schemas/BulkUpsertResult' } } }
        '409': { $ref: '#/components/responses/ConflictIdempotent' }
        default: { $ref: '#/components/responses/Problem' }

  ############################################
  # Portal: menu
  ############################################
  /portal/menu:
    get:
      tags: [Portal Menu]
      summary: List menus
      responses:
        '200':
          description: OK
          content: { application/json: { schema: { $ref: '#/components/schemas/PortalMenuList' } } }
        default: { $ref: '#/components/responses/Problem' }
    post:
      tags: [Portal Menu]
      parameters:
        - $ref: '#/components/parameters/XRequestID'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/PortalMenuCreate' } } }
      responses:
        '201':
          description: Created
          content: { application/json: { schema: { $ref: '#/components/schemas/PortalMenu' } } }
  /portal/menu/{id}:
    patch:
      tags: [Portal Menu]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
        - $ref: '#/components/parameters/XRequestID'
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/PortalMenuUpdate' } } }
      responses:
        '200':
          description: OK
          content: { application/json: { schema: { $ref: '#/components/schemas/PortalMenu' } } }
  /portal/menu/bulk_upsert:
    post:
      tags: [Portal Menu]
      summary: Bulk UPSERT multiple menus
      parameters:
        - $ref: '#/components/parameters/XRequestID'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/BulkPortalMenuRequest' } } }
      responses:
        '200':
          description: OK
          content: { application/json: { schema: { $ref: '#/components/schemas/BulkUpsertResult' } } }

  ############################################
  # Extract (Cron internal)
  ############################################
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    XRequestID:
      name: X-Request-ID
      in: header
      required: false
      schema: { type: string, format: uuid }
      description: Correlation ID per run. Generated by Cron/clients and attached to every request.
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema: { type: string, minLength: 8, maxLength: 256 }
      description: >
        Idempotency key for POST/PUT. Deterministically generated from a time window and payload.
        Matching retries return 200/409 (implementation-dependent) with no side effects.
  responses:
    Problem:
      description: Error
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    ConflictIdempotent:
      description: Treated as replay due to idempotency
      headers:
        Idempotency-Status:
          description: If 'replayed', the request was treated as a replay
          schema: { type: string, enum: [replayed, conflict] }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
  schemas:
    Problem:
      type: object
      properties:
        type: { type: string }
        title: { type: string }
        detail: { type: string }
        status: { type: integer }
      required: [title, status]
...
    CursorList:
      type: object
      properties:
        items: { type: array, items: {} }
        next_cursor: { type: string, nullable: true }

    ########################################
    # Portal model
    ########################################
    PortalModel:
      type: object
      properties:
        id: { type: integer }
        model: { type: string }
        model_table: { type: string }
        label_i18n: { type: object, additionalProperties: true }
        notes: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    PortalModelCreate:
      type: object
      properties:
        model: { type: string }
        model_table: { type: string }
        label_i18n: { type: object, additionalProperties: true }
        notes: { type: string }
      required: [model, model_table]
...
    PortalModelList:
      allOf:
        - $ref: '#/components/schemas/CursorList'
        - type: object
          properties:
            items: { type: array, items: { $ref: '#/components/schemas/PortalModel' } }

    ########################################
    # Portal field
    ########################################
...
    PortalSmartButton:
      type: object
      properties:
        id: { type: integer }
        view_id: { type: integer }
        button_key: { type: string }
        label_i18n: { type: object, additionalProperties: true }
        model: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    PortalSmartButtonCreate:
      type: object
      properties:
        view_id: { type: integer }
        button_key: { type: string }
        label_i18n: { type: object, additionalProperties: true }
        model: { type: string }
      required: [view_id, button_key]
    PortalSmartButtonUpdate:
      type: object
      properties:
        label_i18n: { type: object, additionalProperties: true }
        model: { type: string }
    PortalSmartButtonList:
      allOf:
        - $ref: '#/components/schemas/CursorList'
        - type: object
          properties:
            items: { type: array, items: { $ref: '#/components/schemas/PortalSmartButton' } }
    BulkPortalSmartButtonRequest:
      type: array
      items: { $ref: '#/components/schemas/PortalSmartButtonCreate' }

    ########################################
    # Portal menu
    ########################################
    PortalMenu:
      type: object
      properties:
        id: { type: integer }
        menu_xmlid: { type: string, nullable: true }
        name_ja: { type: string, nullable: true }
        name_en: { type: string, nullable: true }
        name_i18n: { type: object, additionalProperties: true }
        model: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    PortalMenuCreate:
      type: object
      properties:
        menu_xmlid: { type: string }
        name_i18n: { type: object, additionalProperties: true }
        model: { type: string }
    PortalMenuUpdate:
      type: object
      properties:
        name_i18n: { type: object, additionalProperties: true }
        model: { type: string }
    PortalMenuList:
      allOf:
        - $ref: '#/components/schemas/CursorList'
        - type: object
          properties:
            items: { type: array, items: { $ref: '#/components/schemas/PortalMenu' } }
    BulkPortalMenuRequest:
      type: array
      items: { $ref: '#/components/schemas/PortalMenuCreate' }

    ########################################
    # Import / Bootstrap results
    ########################################
    ImportModelRequest:
      type: object
      properties:
        models: { type: array, items: { type: string } }
        scaffold: { type: boolean, default: true }
      required: [models]
...
    ImportResult:
      type: object
      properties:
        created: { type: integer }
        updated: { type: integer }
        skipped: { type: integer }
...